requirements.txt
asgiref==3.8.1
contourpy==1.3.1
cycler==0.12.1
Django==5.1.4
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
drf-standardized-errors==0.14.1
fonttools==4.55.3
kiwisolver==1.4.8
matplotlib==3.9.2
numpy==2.2.1
packaging==24.2
pillow==11.0.0
psycopg2==2.9.10
pydicom==3.0.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
six==1.17.0
sqlparse==0.5.3
tzdata==2024.2
